xcode-list: DRY up the command to find Xcode apps
Instead of repeating this, we can extract a small function out that knows how
to perform the find operation and also passes along any extra flags. This lets
us use a single definition of the find command.

Unfortunately, this breaks shellcheck's ability to parse the flags passed to
find, and so we get some warnings about $1 not being expanded. That's fine,
and we know that, so we'll add a small comment and tell shellcheck to ignore
that warning for the line.